refactor: relocate request-origin helper out of lib into the interface layer
